(#3): Add makefile and fix staticheck issues
